% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{add_fcn}
\alias{add_fcn}
\title{Add function in the existing plot}
\usage{
add_fcn(f, ...)
}
\arguments{
\item{f}{function you want to add}

\item{...}{you can freely use other options from points function}
}
\value{
line function in the plot
}
\description{
Add function in the existing plot
}
