% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{add_line}
\alias{add_line}
\title{Add line in the existing plot}
\usage{
add_line(x, y, ...)
}
\arguments{
\item{x}{x position vector}

\item{y}{y position vector}

\item{...}{you can freely use other options from points function}
}
\value{
line in the plot
}
\description{
Add line in the existing plot
}
