% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{add_points}
\alias{add_points}
\title{Add points in the existing plot}
\usage{
add_points(x, y, ...)
}
\arguments{
\item{x}{x position vector}

\item{y}{y position vector}

\item{...}{you can freely use other options from points function}
}
\value{
points in the plot
}
\description{
Add points in the existing plot
}
