% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{dpareto}
\alias{dpareto}
\title{The probability density function of the pareto distribution.}
\usage{
dpareto(x, alpha, x_m, log.p = FALSE)
}
\arguments{
\item{x}{the point where the pdf to be evaluated}

\item{alpha}{the shape parameter of pareto dist.}

\item{x_m}{the scale parameter of pareto dist.}

\item{log.p}{use log.p = TRUE when you want
to get a result in a log scale}
}
\value{
the pdf value of the pareto dist.
}
\description{
dpareto function evaluates the pdf of pareto dist. at given x with parameters.
}
\examples{
  dpareto(1:5, 2, 3)
  dpareto(1:5, 2, -3:3)
  dpareto(1:5, 2:5, 3)
}
