% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Make your base plot}
\usage{
emptyPlot(xrange = c(0, 1), yrange = c(0, 1), xlab = "", ylab = "", ...)
}
\arguments{
\item{xrange}{the x axis range that you want to show in plot}

\item{yrange}{the y axis range that you want to show in plot}

\item{xlab}{the x axis lable}

\item{ylab}{the y axis lable}

\item{...}{you can freely use other options from plot function}
}
\value{
the plot with coloring and x and y axis line, plus dash support line
}
\description{
emptyPlot simply makes a base for you, when you draw a plot using base package.
It has presetting base theme which is suitable for scientific but pretty.
}
\examples{
  a <- seq(-2, 2, by = 0.1)
  b <- a^2
  emptyPlot(c(-2, 2), c(-2, 2), main = "title", xlab = "x", ylab = "y")
  points(a, b, type = "l", col = "red")
}
