% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mle_gamma}
\alias{mle_gamma}
\title{Maximum Likelihood Estimator for gamma dist.}
\usage{
mle_gamma(sample_g)
}
\arguments{
\item{sample_g}{the given sample to calculate the estimate of the parameters.}
}
\value{
A 2 by 1 vector containing estimate of parameters k and theta: c(k_hat, theta_hat)
}
\description{
mle_gamma function will calculate mle estimate for the shape parameter
"k" and the scale parameter "theta" by using given sample. Since there is no closed form of the mle for gamma distribution, mle_gamma used uniroot function from stats package.
}
\examples{
  my_sample <- rgamma(50, shape = 2, scale = 3)
  mle_gamma(my_sample)
}
