% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mom_gamma}
\alias{mom_gamma}
\title{Method of Moments Estimator for gamma dist.}
\usage{
mom_gamma(sample_g)
}
\arguments{
\item{sample_g}{the given sample to calculate the estimate of the parameters.}
}
\value{
A 2 by 1 vector containing estimate of parameters k and theta: c(k_hat, theta_hat)
}
\description{
mom_gamma function will calculate mom estimate for the shape parameter
"k" and the scale parameter "theta" by using given sample.
}
\examples{
  my_sample <- rgamma(50, shape = 2, scale = 3)
  mom_gamma(my_sample)
}
