% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{newton_method}
\alias{newton_method}
\title{The implementation of newton's method for root finding}
\usage{
newton_method(f, df, x_0, stop_criteria = .Machine$double.eps,
  max_iter = 10^3, left_b = -Inf, right_b = Inf)
}
\arguments{
\item{f}{the function to be applied}

\item{df}{the first derivative of f}

\item{x_0}{the initial point of algorithm}

\item{stop_criteria}{stopping rule}

\item{max_iter}{if the iteration is larger than max_iter, loop will be stopped.}

\item{left_b}{left bound of the solution search}

\item{right_b}{right bound of the solution search}
}
\value{
the root of the function within the given bound
}
\description{
mom_gamma function will calculate mom estimate for the shape parameter
"k" and the scale parameter "theta" by using given sample.
}
\examples{
  x_cu <- function(x){x^3}
  x_sq <- function(x){3*x^2}
  newton_method(x_cu, x_sq, 3)
}
